offer pange done

+++ lib/view/data_source/agent_datasource.dart
@@ -1,3 +1,5 @@
+import 'dart:js_interop';
+
 import 'package:flutter/material.dart';
 import 'package:fuel_delivary_app_admin/common/dialogs/dialogs.dart';
 import 'package:fuel_delivary_app_admin/controller/agent_controller.dart';
@@ -11,10 +13,15 @@ class AgentDataSource extends DataTableSource {
   final AgentController agentController;
   final Function(AgentModel) onEdit;
 
-  AgentDataSource({required this.agentController, required this.agents, required this.onEdit});
+  AgentDataSource(
+      {required this.agentController,
+      required this.agents,
+      required this.onEdit});
 
   @override
-  DataRow? getRow(int index,) {
+  DataRow? getRow(
+    int index,
+  ) {
     // Use the agents list parameter instead of controller's filteredAgents directly
     if (index >= agents.length) return null;
 
@@ -25,7 +32,7 @@ class AgentDataSource extends DataTableSource {
       DataCell(Text(agent.email)),
       DataCell(Text(agent.phone)),
       DataCell(Chip(
-        label: agent.status
+        label: agent.status == false
             ? const Text(
                 "Blocked",
                 style: TextStyle(color: Colors.white),
@@ -34,33 +41,45 @@ class AgentDataSource extends DataTableSource {
                 "Active",
                 style: const TextStyle(color: Colors.white),
               ),
-        color:
-            MaterialStateProperty.all(agent.status ? Colors.red : Colors.green),
+        color: MaterialStateProperty.all(
+            agent.status == false ? Colors.red : Colors.green),
       )),
       DataCell(Row(
         children: [
           IconButton(
             icon: const Icon(Icons.edit),
             onPressed: () {
-             BuildContext? context= navigatorKey.currentContext;
-             if(context!=null){
-              ImageController imageController=ImageController();
+              BuildContext? context = navigatorKey.currentContext;
+              if (context != null) {
+                ImageController imageController = ImageController();
                 List<String> services = [
-    "Car wash",
-    "Deep clean",
-    "Tire change",
-    "EV charge",
-  ];
-              AgentDialogs.addEditAgentDialog(context, imageController: imageController, agentController: agentController, services: services, agent: agent);
-             }
-            
+                  "Car wash",
+                  "Deep clean",
+                  "Tire change",
+                  "EV charge",
+                ];
+                AgentDialogs.addEditAgentDialog(context,
+                    imageController: imageController,
+                    agentController: agentController,
+                    services: services,
+                    agent: agent);
+              }
+
               // Edit action
             },
           ),
           IconButton(
             icon: const Icon(Icons.delete),
-            onPressed: () {
-              agentController.deleteAgent(agent.id!);
+            onPressed: () async {
+              if (navigatorKey.currentContext != null) {
+                bool? delete = await CustomDialogs.delete(
+                    navigatorKey.currentContext!,
+                    "Delete Agent",
+                    "Are you sure you want to delete this agent ?");
+                if (delete != null && delete) {
+                  agentController.deleteAgent(agent.id!);
+                }
+              }
             },
           ),
         ],
